Registration Page Test Case
