+++ src/test/java/Homework16Test.java
@@ -1,0 +1,37 @@
+import org.openqa.selenium.By;
+import org.openqa.selenium.WebDriver;
+import org.openqa.selenium.WebElement;
+import org.openqa.selenium.chrome.ChromeDriver;
+import org.openqa.selenium.chrome.ChromeOptions;
+import org.testng.Assert;
+import org.testng.annotations.Test;
+
+import java.time.Duration;
+public class Homework16Test extends BaseTest {
+
+    @Test
+    public void registrationNavigation() {
+
+        //Pre-Conditions
+        ChromeOptions options = new ChromeOptions();
+        options.addArguments("--remote-allow-origins=*");
+
+        WebDriver driver = new ChromeDriver(options);
+        driver.manage().timeouts().implicitlyWait(Duration.ofSeconds(10));
+
+        //Steps
+        String url = "https://qa.koel.app";
+        driver.get(url);
+
+        //Actions
+        //Step 1
+        WebElement registrationLink = driver.findElement(By.cssSelector("a[href='registration']"));
+        //Step 2
+        registrationLink.click();
+        //Expected Result
+        String registrationUrl = "https://qa.koel.app/registration";
+        Assert.assertEquals(driver.getCurrentUrl(), registrationUrl);
+        //Quit Browser
+        driver.quit();
+    }
+}
