+++ src/test/resources/TestNG.xml
@@ -7,6 +7,8 @@
         <classes>
             <class name="LoginTests"/>
             <class name="BaseTest"/>
+            <class name="Homework16Test"/>
+            <class name="Homework17Test"/>
         </classes>
     </test>
 </suite>
